requirements.txt
python-steam-api
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib